ci: no 7z; also build dictionaries

+++ make_simp_dist.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+BUILD_TYPE="$1"
+
 git archive HEAD -o archive.tar
 mkdir -p dist
 tar xf archive.tar -C dist
@@ -69,11 +71,16 @@ done
 cd ..
 
 # 打包
+
 echo 打包...
+
+if [ x$BUILD_TYPE = x"github" ]; then
+    exit 0
+fi
+
 rm -rf dist/tools
 rm -rf dist/.git
 cp 简体版说明.txt 下载与安装说明.txt 更新纪要.txt dist
-
 
 if [ -x "$(command -v 7zz)" ]; then
     ZIP7=7zz
